Chore refactor IntoParmas and Component
Reorganize code better according to the context
